Rework logic to expect a single signing key
The `verificationMethods` field of a DID document is allowed to contain
multiple entries with the same `id`. The ATProto spec says that only the
first valid entry is used, and any others ignored.

However, DID PLC specifically only allows configuring a single `atproto`
verification method entry, and treats `verificationMethods` as a map in
its JSON formats and Lexicons. It should therefore be impossible to
encounter a DID document with more than one `atproto` entry (and in any
case, the APIs we currently use return the DID PLC structures containing
maps).

Opened did-method-plc/did-method-plc#67 to request clarification of the
specification.